remove-bg/requirements.txt
mcp
rembg[cpu]
pillow
uvicorn
